+++ tests/autofix_files/async100_simple_autofix.py
@@ -1,3 +1,4 @@
+# NOASYNCIO
 # AUTOFIX
 import trio