+++ tests/eval_files/anyio_trio.py
@@ -1,6 +1,9 @@
 # type: ignore
 # ARG --enable=ASYNC220
 # NOTRIO
+# NOASYNCIO
+# set base library so trio doesn't get replaced when running with anyio
+# BASE_LIBRARY anyio
 
 # anyio eval will automatically prepend this test with `--anyio`
 import trio  # isort: skip