+++ tests/eval_files/async102.py
@@ -1,4 +1,6 @@
 # type: ignore
+# NOASYNCIO
+# asyncio has different mechanisms for shielded scopes, so would raise additional errors in this file.
 from contextlib import asynccontextmanager
 
 import trio
@@ -140,7 +142,7 @@ async def foo():
 
 
 # change of functionality, no longer treated as safe
-# https://github.com/Zac-HD/flake8-trio/issues/54
+# https://github.com/python-trio/flake8-async/issues/54
 @asynccontextmanager
 async def foo2():
     try: