+++ tests/eval_files/async102_trio.py
@@ -1,3 +1,4 @@
+# NOASYNCIO
 # NOANYIO - since anyio.Cancelled does not exist
 import trio