+++ tests/eval_files/async104_anyio.py
@@ -1,4 +1,5 @@
 # type: ignore
+# BASE_LIBRARY ANYIO
 import anyio
 
 try: