+++ tests/eval_files/async105_anyio.py
@@ -1,4 +1,6 @@
 # NOTRIO
+# BASE_LIBRARY anyio
+# asyncio obv will not raise any errors on this file
 import anyio