+++ tests/eval_files/async110.py
@@ -1,4 +1,5 @@
 # type: ignore
+# NOASYNCIO
 import trio
 import trio as noerror