+++ tests/eval_files/async111.py
@@ -1,4 +1,5 @@
 # type: ignore
+# NOASYNCIO
 from typing import Any
 
 import trio