+++ tests/eval_files/async116.py
@@ -1,4 +1,5 @@
 # type: ignore
+# NOASYNCIO
 import math
 from math import inf