+++ tests/eval_files/async240.py
@@ -1,4 +1,5 @@
 # type: ignore
+# NOASYNCIO
 import os.path
 from os.path import isfile, normpath, relpath