+++ tests/eval_files/no_library.py
@@ -1,5 +1,6 @@
 # type: ignore
 # ARG --enable=ASYNC220
 # NOANYIO
+# NOASYNCIO
 async def foo():
     subprocess.Popen()  # ASYNC220: 4, 'subprocess.Popen', "trio"